bi_metrics_etl/requirements.txt
pymysql==1.1.0
python-dotenv==1.0.0